+++ dymension/libs/kaspa/demo/user/src/sim/mod.rs
@@ -1,3 +1,6 @@
 mod foo;
-
+mod stats;
+mod util;
+mod round_trip;
+mod demo;
 pub use foo::*;
