+++ dymension/libs/kaspa/demo/user/src/sim/stats.rs
@@ -1,0 +1,4 @@
+
+fn render_stats(stats: Vec<RoundTripStats>, total_spend: u64, total_ops: u64) {}
+
+pub struct RoundTripStats {}
